parser: Fix string literals used as column ids in indexed columns
This fixes indexed columns as e.g. used in primary key constraints which
are quoted using single quotes. Because you would never index a string
literal it is safe to assume that text in single quotes actually is a
column identifier.

See issue #3047.
In theory this breaks cases where they
are actually meant as string literals but as you would never add an
index over a string literal